最新版本

+++ project/common/src/main/java/com/project/domain/Evaluation.java
@@ -7,7 +7,7 @@ import lombok.ToString;
 @Setter
 @ToString
 public class Evaluation {
-    private Integer point; //评分
+    private Double point; //评分
     private String text; // 评价文字
     private String pic; //图片
     private Integer orderId; //所评价的订单号
